refactor: make StatusCodeFuture reusable
