+++ common/src/backends/auth.rs
@@ -28,6 +28,7 @@ use super::{
     cache::{CacheManagement, CacheManager},
     future::ResponseFuture,
     headers::XShuttleAdminSecret,
+    ResponseState, StatusCodeFuture,
 };
 
 pub const EXP_MINUTES: i64 = 5;
@@ -528,51 +529,6 @@ impl<S> Layer<S> for ScopedLayer {
 pub struct Scoped<S> {
     inner: S,
     required: Vec<Scope>,
-}
-
-/// Future for layers that might return a different status code
-#[pin_project]
-pub struct StatusCodeFuture<F> {
-    #[pin]
-    state: ResponseState<F>,
-}
-
-#[pin_project(project = ResponseStateProj)]
-pub enum ResponseState<F> {
-    Called {
-        #[pin]
-        inner: F,
-    },
-    Unauthorized,
-    Forbidden,
-    BadRequest,
-}
-
-impl<F, Error> Future for StatusCodeFuture<F>
-where
-    F: Future<Output = Result<axum::response::Response, Error>>,
-{
-    type Output = Result<axum::response::Response, Error>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let this = self.project();
-
-        match this.state.project() {
-            ResponseStateProj::Called { inner } => inner.poll(cx),
-            ResponseStateProj::Unauthorized => Poll::Ready(Ok(Response::builder()
-                .status(StatusCode::UNAUTHORIZED)
-                .body(Default::default())
-                .unwrap())),
-            ResponseStateProj::Forbidden => Poll::Ready(Ok(Response::builder()
-                .status(StatusCode::FORBIDDEN)
-                .body(Default::default())
-                .unwrap())),
-            ResponseStateProj::BadRequest => Poll::Ready(Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .body(Default::default())
-                .unwrap())),
-        }
-    }
 }
 
 impl<S> Service<Request<Body>> for Scoped<S>
